+++ packages/ethereum-contracts/contracts/interfaces/superfluid/ISuperTokenFactory.sol
@@ -55,7 +55,7 @@ interface ISuperTokenFactory {
      * @param upgradability Upgradability mode
      * @param name Super token name
      * @param symbol Super token symbol
-     * @param adminOverride Admin address override
+     * @param admin Admin address
      * @return superToken The deployed and initialized wrapper super token
      */
     function createERC20Wrapper(
@@ -64,7 +64,7 @@ interface ISuperTokenFactory {
         Upgradability upgradability,
         string calldata name,
         string calldata symbol,
-        address adminOverride
+        address admin
     )
         external
         returns (ISuperToken superToken);
@@ -94,7 +94,7 @@ interface ISuperTokenFactory {
      * @param upgradability Upgradability mode
      * @param name Super token name
      * @param symbol Super token symbol
-     * @param adminOverride Admin address override
+     * @param admin Admin address
      * @return superToken The deployed and initialized wrapper super token
      */
     function createERC20Wrapper(
@@ -102,7 +102,7 @@ interface ISuperTokenFactory {
         Upgradability upgradability,
         string calldata name,
         string calldata symbol,
-        address adminOverride
+        address admin
     )
         external
         returns (ISuperToken superToken);